requirements.txt
aioboto3==10.2.0
aiobotocore==2.4.1
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.6.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.59
botocore==1.27.59
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.1.0
fastapi==0.88.0
faunadb==4.3.1
frozenlist==1.3.3
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
isort==5.11.3
jmespath==1.0.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
multidict==6.0.3
names==0.3.0
nodeenv==1.7.0
orjson==3.8.3
packaging==22.0
platformdirs==2.6.0
pluggy==1.0.0
pydantic==1.10.2
pylint==2.15.9
pyright==1.1.284
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.2
